pyproject.toml
[project]
name = "datasetsmith"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "imagehash>=4.3.2",
    "mcp-agent[openai]>=0.2.5",
    "openai>=1.30.0",
    "pillow>=12.0.0",
    "pydantic>=2.12.4",
    "pymongo>=4.15.4",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "tavily-python>=0.7.13",
    "voyageai>=0.3.5",
]

[project.optional-dependencies]
app = [
    "dagster>=1.12.3",
    "dagster-webserver>=1.12.3",
    "fastapi>=0.121.3",
]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["obp"]
